updating the timer state data
